Optimise TileMap::texture_span
Convert to using fixed-point and skip/combine pixels where possible

Saves ~6ms in blit kart

+++ 32blit/graphics/tilemap.hpp
@@ -38,7 +38,7 @@ namespace blit {
     void draw(Surface *dest, Rect viewport, std::function<Mat3(uint8_t)> scanline_callback = nullptr);
 
   //  void mipmap_texture_span(surface *dest, point s, uint16_t c, vec2 swc, vec2 ewc);
-    void texture_span(Surface *dest, Point s, uint16_t c, Vec2 swc, Vec2 ewc);
+    void texture_span(Surface *dest, Point s, unsigned int c, Vec2 swc, Vec2 ewc);
   };
 
 }
